moved custom update code generation into module
