+++ include/genn/genn/code_generator/generateAll.h
@@ -29,4 +29,7 @@ GENN_EXPORT std::pair<std::vector<std::string>, MemAlloc> generateAll(const Mode
 
 GENN_EXPORT void generateNeuronUpdate(const filesystem::path &outputPath, const ModelSpecMerged &modelMerged, 
                                       const BackendBase &backend, const std::string &suffix = "");
+
+GENN_EXPORT void generateCustomUpdate(const filesystem::path &outputPath, const ModelSpecMerged &modelMerged, 
+                                      const BackendBase &backend, const std::string &suffix = "");
 }
