+++ include/genn/genn/code_generator/groupMerged.h
@@ -1412,124 +1412,6 @@ public:
     static const std::string name;
 };
 
-// ----------------------------------------------------------------------------
-// CustomUpdateGroupMerged
-//----------------------------------------------------------------------------
-class GENN_EXPORT CustomUpdateGroupMerged : public GroupMerged<CustomUpdateInternal>
-{
-public:
-    CustomUpdateGroupMerged(size_t index, const std::string &precision, const std::string &, const BackendBase &backend,
-                            const std::vector<std::reference_wrapper<const CustomUpdateInternal>> &groups);
-
-    //----------------------------------------------------------------------------
-    // Public API
-    //----------------------------------------------------------------------------
-    bool isParamHeterogeneous(size_t index) const;
-    bool isDerivedParamHeterogeneous(size_t index) const;
-
-    void generateRunner(const BackendBase &backend, CodeStream &definitionsInternal,
-                        CodeStream &definitionsInternalFunc, CodeStream &definitionsInternalVar,
-                        CodeStream &runnerVarDecl, CodeStream &runnerMergedStructAlloc) const
-    {
-        generateRunnerBase(backend, definitionsInternal, definitionsInternalFunc, definitionsInternalVar,
-                           runnerVarDecl, runnerMergedStructAlloc, name);
-    }
-
-    boost::uuids::detail::sha1::digest_type getHashDigest() const;
-
-    //----------------------------------------------------------------------------
-    // Static API
-    //----------------------------------------------------------------------------
-    std::string getVarIndex(unsigned int batchSize, VarAccessDuplication varDuplication, const std::string &index) const;
-    std::string getVarRefIndex(bool delay, unsigned int batchSize, VarAccessDuplication varDuplication, const std::string &index) const;
-
-    //----------------------------------------------------------------------------
-    // Static constants
-    //----------------------------------------------------------------------------
-    static const std::string name;
-};
-
-// ----------------------------------------------------------------------------
-// CustomUpdateWUGroupMergedBase
-//----------------------------------------------------------------------------
-class GENN_EXPORT CustomUpdateWUGroupMergedBase : public GroupMerged<CustomUpdateWUInternal>
-{
-public:
-    //----------------------------------------------------------------------------
-    // Public API
-    //----------------------------------------------------------------------------
-    bool isParamHeterogeneous(size_t index) const;
-    bool isDerivedParamHeterogeneous(size_t index) const;
-
-    boost::uuids::detail::sha1::digest_type getHashDigest() const;
-
-    std::string getVarIndex(unsigned int batchSize, VarAccessDuplication varDuplication, const std::string &index) const;
-    std::string getVarRefIndex(unsigned int batchSize, VarAccessDuplication varDuplication, const std::string &index) const;
-
-protected:
-    CustomUpdateWUGroupMergedBase(size_t index, const std::string &precision, const std::string &, const BackendBase &backend,
-                                  const std::vector<std::reference_wrapper<const CustomUpdateWUInternal>> &groups);
-
-};
-
-// ----------------------------------------------------------------------------
-// CustomUpdateWUGroupMerged
-//----------------------------------------------------------------------------
-class GENN_EXPORT CustomUpdateWUGroupMerged : public CustomUpdateWUGroupMergedBase
-{
-public:
-    CustomUpdateWUGroupMerged(size_t index, const std::string &precision, const std::string &timePrecision, const BackendBase &backend,
-                              const std::vector<std::reference_wrapper<const CustomUpdateWUInternal>> &groups)
-    :   CustomUpdateWUGroupMergedBase(index, precision, timePrecision, backend, groups)
-    {
-    }
-
-    //----------------------------------------------------------------------------
-    // Public API
-    //----------------------------------------------------------------------------
-    void generateRunner(const BackendBase &backend, CodeStream &definitionsInternal,
-                        CodeStream &definitionsInternalFunc, CodeStream &definitionsInternalVar,
-                        CodeStream &runnerVarDecl, CodeStream &runnerMergedStructAlloc) const
-    {
-        generateRunnerBase(backend, definitionsInternal, definitionsInternalFunc, definitionsInternalVar,
-                           runnerVarDecl, runnerMergedStructAlloc, name);
-    }
-
-    //----------------------------------------------------------------------------
-    // Static constants
-    //----------------------------------------------------------------------------
-    static const std::string name;
-};
-
-// ----------------------------------------------------------------------------
-// CustomUpdateTransposeWUGroupMerged
-//----------------------------------------------------------------------------
-class GENN_EXPORT CustomUpdateTransposeWUGroupMerged : public CustomUpdateWUGroupMergedBase
-{
-public:
-    CustomUpdateTransposeWUGroupMerged(size_t index, const std::string &precision, const std::string &timePrecision, const BackendBase &backend,
-                                       const std::vector<std::reference_wrapper<const CustomUpdateWUInternal>> &groups)
-    :   CustomUpdateWUGroupMergedBase(index, precision, timePrecision, backend, groups)
-    {
-    }
-
-    //----------------------------------------------------------------------------
-    // Public API
-    //----------------------------------------------------------------------------
-    void generateRunner(const BackendBase &backend, CodeStream &definitionsInternal,
-                        CodeStream &definitionsInternalFunc, CodeStream &definitionsInternalVar,
-                        CodeStream &runnerVarDecl, CodeStream &runnerMergedStructAlloc) const
-    {
-        generateRunnerBase(backend, definitionsInternal, definitionsInternalFunc, definitionsInternalVar,
-                           runnerVarDecl, runnerMergedStructAlloc, name);
-    }
-
-    //----------------------------------------------------------------------------
-    // Static constants
-    //----------------------------------------------------------------------------
-    static const std::string name;
-};
-
 //----------------------------------------------------------------------------
 // CustomUpdateInitGroupMergedBase
 //----------------------------------------------------------------------------
