+++ include/genn/genn/code_generator/modelSpecMerged.h
@@ -11,6 +11,7 @@
 
 // GeNN code generator includes
 #include "code_generator/groupMerged.h"
+#include "code_generator/customUpdateGroupMerged.h"
 #include "code_generator/neuronUpdateGroupMerged.h"
 #include "code_generator/supportCodeMerged.h"
