+++ src/genn/genn/genn.vcxproj
@@ -23,8 +23,8 @@
     <ClCompile Include="code_generator\backendSIMT.cc" />
     <ClCompile Include="code_generator\codeGenUtils.cc" />
     <ClCompile Include="code_generator\codeStream.cc" />
+    <ClCompile Include="code_generator\customUpdateGroupMerged.cc" />
     <ClCompile Include="code_generator\generateAll.cc" />
-    <ClCompile Include="code_generator\generateCustomUpdate.cc" />
     <ClCompile Include="code_generator\generateInit.cc" />
     <ClCompile Include="code_generator\generateMakefile.cc" />
     <ClCompile Include="code_generator\generateMSBuild.cc" />
@@ -63,8 +63,8 @@
     <ClInclude Include="..\..\..\include\genn\genn\code_generator\backendSIMT.h" />
     <ClInclude Include="..\..\..\include\genn\genn\code_generator\codeGenUtils.h" />
     <ClInclude Include="..\..\..\include\genn\genn\code_generator\codeStream.h" />
+    <ClInclude Include="..\..\..\include\genn\genn\code_generator\customUpdateGroupMerged.h" />
     <ClInclude Include="..\..\..\include\genn\genn\code_generator\generateAll.h" />
-    <ClInclude Include="..\..\..\include\genn\genn\code_generator\generateCustomUpdate.h" />
     <ClInclude Include="..\..\..\include\genn\genn\code_generator\generateInit.h" />
     <ClInclude Include="..\..\..\include\genn\genn\code_generator\generateMakefile.h" />
     <ClInclude Include="..\..\..\include\genn\genn\code_generator\generateMSBuild.h" />
